requirements.txt
cv2
tkinter
Pillow
selenium
requests
hashlib
cmake
dlib
face_recognition
imutils